07_django_paginations_filter_search/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
Faker==19.6.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
